pyproject.toml
[project]
name = "mognoapi"
version = "0.1.0"
description = "An etl agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "boto3>=1.38.15",
    "botocore>=1.38.15",
    "certifi>=2025.4.26",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.2",
    "cramjam>=2.10.0",
    "cryptography>=44.0.3",
    "dnspython>=2.7.0",
    "faker>=37.1.0",
    "h11>=0.16.0",
    "h2>=4.2.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "jmespath>=1.0.1",
    "packaging>=25.0",
    "psycopg2-binary>=2.9.10",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.22",
    "pymongo>=4.12.1",
    "pymongo-auth-aws>=1.3.0",
    "pymongocrypt>=1.14.0",
    "pyopenssl>=25.0.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-snappy>=0.7.3",
    "requests>=2.32.3",
    "s3transfer>=0.12.0",
    "service-identity>=24.2.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "winkerberos>=0.12.2",
    "zstandard>=0.23.0",
]
